Enable the building of Scala 2.12 & 2.13 variants of module 'shared'
