+++ integration/spark/gradle.properties
@@ -2,3 +2,6 @@ jdk8.build=true
 version=1.9.0-SNAPSHOT
 spark.version=3.5.0
 org.gradle.jvmargs=-Xmx1G
+
+scala.binary.version=2.12
+shared.spark.version=3.2.4
